+++ src/main/java/com/unstoppabledomains/resolution/contracts/JsonProvider.java
@@ -1,15 +1,21 @@
 package com.unstoppabledomains.resolution.contracts;
 
+import com.google.gson.FieldNamingPolicy;
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.google.gson.JsonObject;
 import java.io.IOException;
 import java.net.HttpURLConnection;
 
 public class JsonProvider extends DefaultProvider {
   private String method = "GET";
+  private Gson gson;
   
   public JsonProvider() {
     super();
+    gson = new GsonBuilder()
+      .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
+      .create();
   }
   
   public String getMethod() {
@@ -22,7 +28,7 @@ public class JsonProvider extends DefaultProvider {
 
   public <T> T request(String url, JsonObject body, java.lang.Class<T> classOfT) throws IOException {
     String rawResponse = super.rawRequest(url, body);
-    return new Gson().fromJson(rawResponse, classOfT);
+    return gson.fromJson(rawResponse, classOfT);
   }
   
   @Override
